supply side histogram xlim
